chore: clean test files

+++ contracts/6/protocol/interfaces/IVault.sol
@@ -14,18 +14,18 @@ Alpaca Fin Corporation
 pragma solidity 0.6.6;
 
 interface IVault {
-
-  /// @dev Return the total ERC20 entitled to the token holders. Be careful of unaccrued interests.
+  /// @notice Return the total ERC20 entitled to the token holders. Be careful of unaccrued interests.
   function totalToken() external view returns (uint256);
 
-  /// @dev Add more ERC20 to the bank. Hope to get some good returns.
+  /// @notice Add more ERC20 to the bank. Hope to get some good returns.
   function deposit(uint256 amountToken) external payable;
 
-  /// @dev Withdraw ERC20 from the bank by burning the share tokens.
+  /// @notice Withdraw ERC20 from the bank by burning the share tokens.
   function withdraw(uint256 share) external;
 
-  /// @dev Request funds from user through Vault
-  function requestFunds(address targetedToken, uint amount) external;
+  /// @notice Request funds from user through Vault
+  function requestFunds(address targetedToken, uint256 amount) external;
 
+  /// @notice Underlying token address
   function token() external view returns (address);
 }
